Basic functionality to display icon
Basic functionality of getting the current weather for a given location
(currently hard-coded to the coordinates for Ramsey Solutions) and
displaying the icon for the current weather type. A simple unit test is
included that tests the WeatherManager's ability to get the current
conditions.

Still need to implement logic for keeping the icon cache below the
maximum size.