Get the user's current location and UI updates
Update project to allow getting the user's location when the app is
foreground and update ViewController to request and handle the
location. Add a loading indicator and logic to display/hide it. Also
add a refresh button in the navigation bar. Add attribution to
OpenWeatherMap on the launch screen as well along with some project
group clean-up.